support negative index

+++ src/common/context/ExpressionContext.h
@@ -62,7 +62,7 @@ public:
     virtual Value getEdge() const = 0;
 
     // Get Value by Column number
-    virtual Value getColumn(size_t index) const = 0;
+    virtual Value getColumn(int32_t index) const = 0;
 
     virtual void setVar(const std::string& var, Value val) = 0;
 };
