+++ src/common/expression/test/ExpressionContextMock.h
@@ -117,7 +117,7 @@ public:
         return Value();
     }
 
-    Value getColumn(size_t index) const override;
+    Value getColumn(int32_t index) const override;
 
     void setVar(const std::string& var, Value val) override {
         UNUSED(var);
